requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
Django==5.1.4
djangorestframework==3.15.2
executing==2.1.0
gunicorn==23.0.0
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
